Allow installing previous versions

+++ install.sh
@@ -296,8 +296,10 @@ unpack() {
   if [ "$local_install" -ne '1' ]; then
     $remove_command "$pkg_name"
     dir_name=$(echo "${pkg_name}" | sed -E -e 's/(.*)(\.tar\.gz|\.zip)/\1/')
-    mv -f "${output_dir}/${dir_name}/"* "${output_dir}"
-    rmdir "${output_dir}/${dir_name}"
+    if [ -d "${output_dir}/${dir_name}/"]; then
+      mv -f "${output_dir}/${dir_name}/"* "${output_dir}"
+      rmdir "${output_dir}/${dir_name}"
+    fi
   fi
 
   log "Package has been unpacked successfully"
